refactor: extract collectMetrics and add schema memoization
Extract the metric-collection pipeline from handleScore into a
standalone collect-metrics.ts module with two exports:
- collectMetrics(): low-level function used by handleScore
- collectDocumentMetrics(): high-level convenience used by tests

This eliminates ~300 lines of duplicated walker setup across three
test files (document-metrics, example-coverage, index) and ensures
tests exercise the same code path as the production command.

Add $ref-keyed memoization to walkSchema so repeated references to
the same component schema return cached stats instead of re-walking.
Stripe API: 37.6s → 11.3s (3.3× faster).

Made-with: Cursor
